moving EnvDte use to higher level projects
