+++ Source/nHydrate.Generator.Common/Util/Constants.cs
@@ -1,0 +1,10 @@
+namespace nHydrate.Generator.Common.Util
+{
+    public enum FileStateConstants
+    {
+        Success,
+        Skipped,
+        Failed,
+    }
+
+}
